refactor(ndt): move spec into its own package
